==> 2_scraping_restaurants.ipynb <==
**SCRAPE POSTCODES FOR TOP RESTAURANTS FROM THE GOOD FOOD GUIDE**


import pandas as pd
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
import requests
from tqdm import tqdm
import time


names = []
postcodes = []

for x in tqdm(range(1, 45)):
    url = 'https://www.thegoodfoodguide.co.uk/search/q/London/restaurant-page/{}/'.format(x)
    r = requests.get(url)
    soup = BeautifulSoup(r.text, 'html.parser')
    time.sleep(np.abs(np.random.normal(loc = 3, scale = 0.5)))
    
    for restaurant in soup.find_all('a', attrs={'class': 'o-faux-block-link__overlay'}):
        href = restaurant['href']
        try:
            new_url = 'https://www.thegoodfoodguide.co.uk{}'.format(href)
            r = requests.get(new_url)
            soup = BeautifulSoup(r.text, 'html.parser')
            postcode = soup.find('span', attrs={'itemprop': 'postalCode'}).text.strip()
            postcodes.append(postcode)
        except:
            postcodes.append(np.nan)
            
        try:
            new_url = 'https://www.thegoodfoodguide.co.uk{}'.format(href)
            r = requests.get(new_url)
            soup = BeautifulSoup(r.text, 'html.parser')
            name = soup.find('span', attrs={'itemprop': 'streetAddress'}).text.strip()
            names.append(name)
        except:
            names.append(np.nan)
        
        

restaurants = pd.DataFrame({'name': names,
                     'postcode': postcodes})

restaurants = restaurants.dropna().reset_index(drop = True)

restaurants['name'] = restaurants['name'].str.strip(',')

restaurants.to_csv(r'/Users/davidrhode/Desktop/GA/DSI15-lessons/project/project-capstone/Restaurants/good-restaurants.csv', index=False)
