==> 3_scraping_estate_agents.ipynb <==
**SCRAPE ALL IN LONDON FOR ESTATE AGENTS POSTCODES**

import pandas as pd
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
import requests
from tqdm import tqdm
import time


names = []
addresses = []

for x in tqdm(range(1, 294)):
    url = 'https://www.allinlondon.co.uk/directory/estate-agents?page={}'.format(x)
    r = requests.get(url)
    soup = BeautifulSoup(r.text, 'html.parser')
    time.sleep(np.abs(np.random.normal(loc = 3, scale = 0.5)))
    
    for agent in soup.find_all('div', attrs={'class':'bg-brand-section p-3 mb-3'}):
        try:
            names.append(agent.find('a', attrs={'class':'h5'}).text.strip())
        except:
            names.append(np.nan)

        try:
            addresses.append(agent.find('div', attrs={'class':'text-muted mb-1'}).text.strip())
        except:
            addresses.append(np.nan)


estate_agents = pd.DataFrame({'name': names,
                     'address': addresses})

estate_agents = estate_agents.dropna().reset_index(drop = True)
estate_agents

# use a regular expression to extract the postcode from the address column

import re
pattern_1 = r'([A-Z][A-Z0-9]{0,1}[A-Z0-9]{0,2} {0,}[0-9][A-Z]{0,2})'

estate_agents['postcode'] = estate_agents['address'].str.extract(pattern_1)
estate_agents.head()

estate_agents = estate_agents.drop(columns = ['address'])
estate_agents.head()

estate_agents.to_csv(r'/Users/davidrhode/Desktop/GA/DSI15-lessons/project/project-capstone/Estate Agents/estate_agents.csv', index=False)
