==> 4_scraping_florists.ipynb <==
**SCRAPE ALL IN LONDON FOR FLORISTS POSTCODES**


import pandas as pd
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
import requests
from tqdm import tqdm
import time


names = []
addresses = []

for x in tqdm(range(1, 52)):
    url = 'https://www.allinlondon.co.uk/directory/florists?page={}'.format(x)
    r = requests.get(url)
    soup = BeautifulSoup(r.text, 'html.parser')
    time.sleep(np.abs(np.random.normal(loc = 3, scale = 0.5)))
    
    for florist in soup.find_all('div', attrs={'class':'bg-brand-section p-3 mb-3'}):
        try:
            names.append(florist.find('a', attrs={'class':'h5'}).text.strip())
        except:
            names.append(np.nan)

        try:
            addresses.append(florist.find('div', attrs={'class':'text-muted mb-1'}).text.strip())
        except:
            addresses.append(np.nan)

florists = pd.DataFrame({'name': names, 'address': addresses})
florists = florists.dropna().reset_index(drop = True)
florists

# use a regular expression to extract the postcode from the address column

import re
pattern_1 = r'([A-Z][A-Z0-9]{0,1}[A-Z0-9]{0,2} {0,}[0-9][A-Z]{0,2})'

florists['postcode'] = florists['address'].str.extract(pattern_1)
florists.head()

florists = florists.drop(columns = ['address'])
florists.head()

florists.to_csv(r'/Users/davidrhode/Desktop/GA/DSI15-lessons/project/project-capstone/Florists/florists.csv', index=False)
