==> 5_scraping_delicatessens.ipynb <==
**SCRAPE ALL IN LONDON FOR DELICATESSEN POSTCODES**

import pandas as pd
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
import requests
from tqdm import tqdm
import time


names = []
addresses = []

for x in tqdm(range(1, 64)):
    url = 'https://www.allinlondon.co.uk/directory/delicatessens?page={}'.format(x)
    r = requests.get(url)
    soup = BeautifulSoup(r.text, 'html.parser')
    time.sleep(np.abs(np.random.normal(loc = 3, scale = 0.5)))
    
    for deli in soup.find_all('div', attrs={'class':'bg-brand-section p-3 mb-3'}):
        try:
            names.append(deli.find('a', attrs={'class':'h5'}).text.strip())
        except:
            names.append(np.nan)

        try:
            addresses.append(deli.find('div', attrs={'class':'text-muted mb-1'}).text.strip())
        except:
            addresses.append(np.nan)



delicatessens = pd.DataFrame({'name': names,
                     'address': addresses})

delicatessens = delicatessens.dropna()

# use a regular expression to extract the postcode from the address column

import re
pattern_1 = r'([A-Z][A-Z0-9]{0,1}[A-Z0-9]{0,2} {0,}[0-9][A-Z]{0,2})'

delicatessens['postcode'] = delicatessens['address'].str.extract(pattern_1)
delicatessens.head()

delicatessens = delicatessens.drop(columns = ['address'])

delicatessens = delicatessens.reset_index(drop = True)
delicatessens

delicatessens.to_csv(r'/Users/davidrhode/Desktop/GA/DSI15-lessons/project/project-capstone/Delicatessens/delicatessens.csv', index=False)
