==> 7_processing_crime.ipynb <==
** PROCESS CRIME DATA FROM CITY OF LONDON POLICE AND METROPOLITAN POLICE**

Import data, clean it up and check that there are no underlying trends that we need to be aware of.


# import crime data from City of London Police going back to early 2018, including geocoordinates for each crime

import pandas as pd

cl_1 = pd.read_csv('2018-01-city-of-london-street.csv')
cl_2 = pd.read_csv('2018-02-city-of-london-street.csv')
cl_3 = pd.read_csv('2018-03-city-of-london-street.csv')
cl_4 = pd.read_csv('2018-04-city-of-london-street.csv')
cl_5 = pd.read_csv('2018-05-city-of-london-street.csv')
cl_6 = pd.read_csv('2018-06-city-of-london-street.csv')
cl_7 = pd.read_csv('2018-07-city-of-london-street.csv')
cl_8 = pd.read_csv('2018-08-city-of-london-street.csv')
cl_9 = pd.read_csv('2018-09-city-of-london-street.csv')
cl_10 = pd.read_csv('2018-10-city-of-london-street.csv')
cl_11 = pd.read_csv('2018-11-city-of-london-street.csv')
cl_12 = pd.read_csv('2018-12-city-of-london-street.csv')
cl_13 = pd.read_csv('2019-01-city-of-london-street.csv')
cl_14 = pd.read_csv('2019-02-city-of-london-street.csv')
cl_15 = pd.read_csv('2019-03-city-of-london-street.csv')
cl_16 = pd.read_csv('2019-04-city-of-london-street.csv')
cl_17 = pd.read_csv('2019-05-city-of-london-street.csv')
cl_18 = pd.read_csv('2019-06-city-of-london-street.csv')
cl_19 = pd.read_csv('2019-07-city-of-london-street.csv')
cl_20 = pd.read_csv('2019-08-city-of-london-street.csv')
cl_21 = pd.read_csv('2019-09-city-of-london-street.csv')
cl_22 = pd.read_csv('2019-10-city-of-london-street.csv')
cl_23 = pd.read_csv('2019-11-city-of-london-street.csv')
cl_24 = pd.read_csv('2019-12-city-of-london-street.csv')
cl_25 = pd.read_csv('2020-01-city-of-london-street.csv')
cl_26 = pd.read_csv('2020-02-city-of-london-street.csv')
cl_27 = pd.read_csv('2020-03-city-of-london-street.csv')
cl_crime = pd.concat([cl_1, cl_2, cl_3, cl_4, cl_5, cl_6, cl_7, cl_8, cl_9, cl_10, cl_11, cl_12, cl_13, cl_14, cl_15, cl_16, cl_17, cl_18, cl_19, cl_20, cl_21, cl_22, cl_23, cl_24, cl_25, cl_26, cl_27])

# import crime data from Metropolitan Police going back to early 2018, including geocoordinates for each crime


met_1 = pd.read_csv('2018-01-metropolitan-street.csv')
met_2 = pd.read_csv('2018-02-metropolitan-street.csv')
met_3 = pd.read_csv('2018-03-metropolitan-street.csv')
met_4 = pd.read_csv('2018-04-metropolitan-street.csv')
met_5 = pd.read_csv('2018-05-metropolitan-street.csv')
met_6 = pd.read_csv('2018-07-metropolitan-street.csv')
met_7 = pd.read_csv('2018-07-metropolitan-street.csv')
met_8 = pd.read_csv('2018-08-metropolitan-street.csv')
met_9 = pd.read_csv('2018-09-metropolitan-street.csv')
met_10 = pd.read_csv('2018-10-metropolitan-street.csv')
met_11 = pd.read_csv('2018-11-metropolitan-street.csv')
met_12 = pd.read_csv('2018-12-metropolitan-street.csv')
met_13 = pd.read_csv('2019-01-metropolitan-street.csv')
met_14 = pd.read_csv('2019-02-metropolitan-street.csv')
met_15 = pd.read_csv('2019-03-metropolitan-street.csv')
met_16 = pd.read_csv('2019-04-metropolitan-street.csv')
met_17 = pd.read_csv('2019-05-metropolitan-street.csv')
met_18 = pd.read_csv('2019-06-metropolitan-street.csv')
met_19 = pd.read_csv('2019-07-metropolitan-street.csv')
met_20 = pd.read_csv('2019-08-metropolitan-street.csv')
met_21 = pd.read_csv('2019-09-metropolitan-street.csv')
met_22 = pd.read_csv('2019-10-metropolitan-street.csv')
met_23 = pd.read_csv('2019-11-metropolitan-street.csv')
met_24 = pd.read_csv('2019-12-metropolitan-street.csv')
met_25 = pd.read_csv('2020-01-metropolitan-street.csv')
met_26 = pd.read_csv('2020-02-metropolitan-street.csv')
met_27 = pd.read_csv('2020-03-metropolitan-street.csv')
met_crime = pd.concat([met_1, met_2, met_3, met_4, met_5, met_6, met_7, met_8, met_9, met_10, met_11, met_12, met_13, met_14, met_15, met_16, met_17, met_18, met_19, met_20, met_21, met_22, met_23, met_24, met_25, met_26, met_27])

# concatenate City of London and Metropolitan Police data, select relevant columns

all_crime = pd.concat([met_crime, cl_crime])
cols = ['Month', 'Longitude', 'Latitude', 'LSOA code', 'Crime type']
all_crime = all_crime[cols]

# There are about 30k crimes (out of 2.5 million) with no location information at all, these will need to be dropped

all_crime = all_crime.dropna()

all_crime['Crime type'].value_counts()

# process the date data to allow us to check for underlying trends
# note - information for 2020 covers only the first three months of the year

all_crime['Month'] = pd.to_datetime(all_crime['Month'])
all_crime = all_crime.rename(columns = {'Month': 'date'})
all_crime['year'] = all_crime['date'].dt.year
all_crime['month'] = all_crime['date'].dt.month
all_crime.groupby('year').count()

# confirm there is no strong underlying trend when we break data out by crime type

all_crime.groupby(['Crime type','year']).count()

# copy dataframe and filter columns

all_crime = all_crime.rename(columns = {'LSOA code': 'lsoa'})
all_crime_copy_2 = all_crime.copy()
cols = ['date', 'lsoa']
all_crime_copy_2 = all_crime_copy_2[cols]

# group the data to confirm how much crime was committed in each LSOA region during the period in question

grouped_crime = all_crime_copy_2.groupby('lsoa').count().reset_index()
grouped_crime = grouped_crime.rename(columns = {'date': 'crime_count_lsoa'})

grouped_crime.to_csv('../../../big_files/grouped_crime.csv')
