==> 8_processing_perception_crime.ipynb <==
**PERCEPTION OF CRIME SURVEYS**

Import data, clean it up, check there are no underlying trends that we need to be aware of.


#import crime surveys by Mayor of London's office, concatenate into single dataframe

import pandas as pd
import geopandas as gpd

q2_2017_18 = pd.read_csv('Q2 2017-18.csv')
q3_2017_18 = pd.read_csv('Q3 2017-18.csv')
q1_2017_18 = pd.read_csv('Q1 2017-18.csv')
q3_2016_17 = pd.read_csv('Q3 2016-17.csv')
q2_2016_17 = pd.read_csv('Q2 2016-17.csv')
q4_2016_17 = pd.read_csv('Q4 2016-17.csv')
q1_2018_19 = pd.read_csv('Q1 2018-19.csv')
q4_2017_18 = pd.read_csv('Q4 2017-18.csv')
q2_2018_19 = pd.read_csv('Q2 2018-19.csv')
q3_2018_19 = pd.read_csv('Q3 2018-19.csv')
q2_2017_18['date'] = 'q2_2017_18'
q3_2017_18['date'] = 'q3_2017_18'
q1_2017_18['date'] = 'q1_2017_18'
q3_2016_17['date'] = 'q3_2016_17'
q4_2017_18['date'] = 'q4_2017_18'
q2_2016_17['date'] = 'q2_2016_17'
q4_2016_17['date'] = 'q4_2016_17'
q1_2018_19['date'] = 'q1_2018_19'
q4_2017_18['date'] = 'q4_2017_18'
q2_2018_19['date'] = 'q2_2018_19'
q3_2018_19['date'] = 'q3_2018_19'

all_crime_perception = pd.concat([q2_2017_18, q3_2017_18, q1_2017_18, q3_2016_17, q2_2016_17, q4_2016_17, q1_2018_19, q4_2017_18, q2_2018_19, q3_2018_19])
all_crime_perception = all_crime_perception.drop_duplicates()


all_crime_perception.head()

#filter for relevant columns, drop a handful of missing values

all_crime_perception = all_crime_perception.drop(columns = ['Unnamed: 22', 'Total Notifiable Offences rate', 'The police in this area understand the issues that affect this community']) 
all_crime_perception = all_crime_perception.reset_index(drop = False)
all_crime_perception = all_crime_perception.dropna()
all_crime_perception = all_crime_perception.drop([0])

all_crime_perception[['Taking everything into account how good a job do you think the police in your area are doing?',
       'The police are dealing with the things that matter to this community',
       'The police can be relied upon to be there when needed',
       'The police treat everyone fairly regardless of who they are',
       'The police listen to the concerns of local people',
       'I feel well informed about local police activities over the last 12 months',
       'How to contact their local SNT or ward officer',
       'The police would treat you with respect if you had contact with them for any reason',
       'How well do you think the police tackle gun crime?',
       'How well do you think the police tackle drug dealing and drug use?',
       'Are gangs a problem in this area?',
       'Are you worried about ASB in this area?',
       'Are you worried about crime in this area?',
       'How safe do you feel walking alone in this area after dark?',
       'How well do you think the Metropolitan Police provide a visible patrolling presence',
       'Are you satisfied with your life as a whole nowadays?',
       'The police in this area are friendly and approachable']] = all_crime_perception[['Taking everything into account how good a job do you think the police in your area are doing?',
       'The police are dealing with the things that matter to this community',
       'The police can be relied upon to be there when needed',
       'The police treat everyone fairly regardless of who they are',
       'The police listen to the concerns of local people',
       'I feel well informed about local police activities over the last 12 months',
       'How to contact their local SNT or ward officer',
       'The police would treat you with respect if you had contact with them for any reason',
       'How well do you think the police tackle gun crime?',
       'How well do you think the police tackle drug dealing and drug use?',
       'Are gangs a problem in this area?',
       'Are you worried about ASB in this area?',
       'Are you worried about crime in this area?',
       'How safe do you feel walking alone in this area after dark?',
       'How well do you think the Metropolitan Police provide a visible patrolling presence',
       'Are you satisfied with your life as a whole nowadays?',
       'The police in this area are friendly and approachable']].astype(float)

# check that attitudes haven't been subject to major underlying trends during the period in question

all_crime_perception.groupby('date').agg(['mean'])

# copy dataframe and filter columns

perceptions = all_crime_perceptions.copy()
cols = ['Neighbourhood','Are gangs a problem in this area?', 'Are you worried about crime in this area?', 'How safe do you feel walking alone in this area after dark?', 'Are you satisfied with your life as a whole nowadays?']
perceptions = perceptions[cols]


perceptions.to_csv('../../../../big_files/perception_crime.csv')

# clean up the 'Neighbourhood' column

nh = perceptions['Neighbourhood'].apply(lambda x: x.split('_')[0])
nh = nh.apply(lambda x: x.split('-')[0])
nh = nh.str.strip()
nh = nh.replace('HammersmithandFulham', 'Hammersmith and Fulham')
nh = nh.replace('KingstonuponThames', 'Kingston upon Thames')
nh = nh.replace('KensingtonandChelsea', 'Kensington and Chelsea')
nh = nh.replace('RichmonduponThames', 'Richmond upon Thames')
nh = nh.replace('TowerHamlets', 'Tower Hamlets')
nh = nh.replace('WalthamForest', 'Waltham Forest')
nh = nh.replace('BarkingandDagenham', 'Barking and Dagenham')
perceptions['Neighbourhood'] = nh
perceptions = perceptions.rename(columns = {'Neighbourhood': 'borough_name'})

# groupby borough to get perception data for each region

borough_perceptions = perceptions.groupby('borough_name')['Are gangs a problem in this area?', 'Are you worried about crime in this area?', 'How safe do you feel walking alone in this area after dark?', 'Are you satisfied with your life as a whole nowadays?'].mean().reset_index()
