==> 9_processing_tubes.ipynb <==
**PROCESSING DATA FOR TUBE STOPS**

Create multiple geo-dataframes based on different radius around each tube stop (50m, 200m etc)


import pandas as pd
import geopandas as gpd

#import geodata for stops on each of the tube lines, concatenate and drop duplicates (stops that are on more than one line)

bakerloo = pd.read_csv('Bakerloo.csv')
central = pd.read_csv('Central .csv')
circle = pd.read_csv('Circle.csv')
district = pd.read_csv('District.csv')
hammersmith = pd.read_csv('Hammersmith and City.csv')
jubilee = pd.read_csv('Jubilee.csv')
metropolitan = pd.read_csv('Metropolitan.csv')
northern = pd.read_csv('Northern.csv')
piccadilly = pd.read_csv('Piccadilly.csv')
victoria = pd.read_csv('Victoria.csv')

tubes = pd.concat([victoria, piccadilly, northern, metropolitan, jubilee, hammersmith, district, circle, central, bakerloo])
tubes = tubes.drop_duplicates()

# convert to geo-dataframe, set CRS to cartesian map coordinates

tubes_gdf = gpd.GeoDataFrame(tubes, geometry = gpd.points_from_xy(tubes['OS X'], tubes['OS Y']))
tubes_gdf = tubes_gdf.set_crs(epsg= 27700)
tubes_gdf = tubes_gdf.to_crs(epsg= 27700)
tubes_gdf.crs

# create a 50 metre radius around each tube station

tubes_50m = tubes_gdf.copy()
tubes_50m['geometry'] = tubes_50m['geometry'].buffer(50)
cols = ['Zone', 'geometry']
tubes_50m = tubes_50m[cols]
tubes_50m = tubes_50m.rename(columns = {'Zone': 'zone_50m'})
tubes_50m.head(1)

# create a 200 metre radius around each tube station

tubes_200m = tubes_gdf.copy()
tubes_200m['geometry'] = tubes_200m['geometry'].buffer(200)
tubes_200m = tubes_200m[cols]
tubes_200m = tubes_200m.rename(columns = {'Zone': 'zone_200m'})
tubes_200m.head(1)

# create a 400 metre radius around each tube station

tubes_400m = tubes_gdf.copy()
tubes_400m['geometry'] = tubes_400m['geometry'].buffer(400)
tubes_400m = tubes_400m[cols]
tubes_400m = tubes_400m.rename(columns = {'Zone': 'zone_400m'})
tubes_400m.head(1)

# create a 800 metre radius around each tube station

tubes_800m = tubes_gdf.copy()
tubes_800m['geometry'] = tubes_800m['geometry'].buffer(800)
tubes_800m = tubes_800m[cols]
tubes_800m = tubes_800m.rename(columns = {'Zone': 'zone_800m'})
tubes_800m.head(1)

# create a 1600 metre radius around each tube station

tubes_1600m = tubes_gdf.copy()
tubes_1600m['geometry'] = tubes_1600m['geometry'].buffer(1600)
tubes_1600m = tubes_1600m[cols]
tubes_1600m = tubes_1600m.rename(columns = {'Zone': 'zone_1600m'})
tubes_1600m.head(1)

# create a 2400 metre radius around each tube station

tubes_2400m = tubes_gdf.copy()
tubes_2400m['geometry'] = tubes_2400m['geometry'].buffer(2400)
tubes_2400m = tubes_2400m[cols]
tubes_2400m = tubes_2400m.rename(columns = {'Zone': 'zone_2400m'})
tubes_2400m.head(1)

tubes_2400m.to_file('../../../big_files/tubes2400m.shp')
tubes_1600m.to_file('../../../big_files/tubes1600m.shp')
tubes_800m.to_file('../../../big_files/tubes800m.shp')
tubes_400m.to_file('../../../big_files/tubes400m.shp')
tubes_200m.to_file('../../../big_files/tubes200m.shp')
tubes_50m.to_file('../../../big_files/tubes50m.shp')
