==> 10_knife_gun.ipynb <==
**GUN AND KNIFE CRIME DATA**

Only available on rolling twelve month cycle

Filter for 'ward_code' and total crime volumes

import geopandas as gpd
import pandas as pd


gun = pd.read_csv('../../../big_files/gun.csv')
gun = gun.loc[:, ['ward_code', 'total_gun_crimes']]
knife = pd.read_csv('../../../big_files/knife.csv')
knife = knife.loc[:, ['ward_code', 'total_knife_crimes']]
