==> 11_deprivation_data.ipynb <==
**DEPARTMENT OF HOUSING DEPRIVATION DATA FOR 2019**

The seven domains of deprivation are as follows:

The Income Deprivation Domain measures the proportion of the population experiencing deprivation relating to low income. The definition of low income used includes both those people that are out-of-work, and those that are in work but who have low earnings (and who satisfy the respective means tests).

The Employment Deprivation Domain measures the proportion of the working-age population in an area involuntarily excluded from the labour market. This includes people who would like to work but are unable to do so due to unemployment, sickness or disability, or caring responsibilities.

The Education, Skills and Training Deprivation Domain measures the lack of attainment and skills in the local population. The indicators fall into two sub-domains: one relating to children and young people and one relating to adult skills. 

The Health Deprivation and Disability Domain measures the risk of premature death and the impairment of quality of life through poor physical or mental health. The domain measures morbidity, disability and premature mortality but not aspects of behaviour or environment that may be predictive of future health deprivation.

The Crime Domain measures the risk of personal and material victimisation at local level. 

The Barriers to Housing and Services Domain measures the physical and financial accessibility of housing and local services. The indicators fall into two sub-domains: ‘geographical barriers’, which relate to the physical proximity of local services, and ‘wider barriers’ which includes issues relating to access to housing such as affordability.

The Living Environment Deprivation Domain measures the quality of the local environment. The indicators fall into two sub-domains. The ‘indoors’ living environment measures the quality of housing; while the ‘outdoors’ living environment contains measures of air quality and road traffic accidents.

LSOAs (Lower-layer Super Output Areas) are small areas designed to be of a similar population size, with an average of approximately 1,500 residents or 650 households. There are 32,844 Lower-layer Super Output Areas (LSOAs) in England. They were produced by the Office for National Statistics for the reporting of small area statistics.


#import data and filter for relevant columns

import pandas as pd
import geopandas as gpd
deprivation = pd.read_csv('../../../big_files/deprivation.csv')
cols = ['LSOA code (2011)','Income Rank (where 1 is most deprived)', 'Employment Rank (where 1 is most deprived)', 'Education, Skills and Training Rank (where 1 is most deprived)', 'Health Deprivation and Disability Score', 'Crime Rank (where 1 is most deprived)', 'Barriers to Housing and Services Rank (where 1 is most deprived)', 'Living Environment Score']
deprivation = deprivation.rename(columns = {'LSOA code (2011)': 'lsoa'})


#import data for free school meal eligibility, no need to filter at this stage

school_meals = pd.read_csv('LSOA_FreeSchoolMeals_London_LDS2015.csv')
