==> 12_centrality.ipynb <==
**CREATE MEASURE OF CENTRALITY**

Trafalgar Square / Charing Cross is generally seen as most central point in London.

We can create a dataframe containing its coordinates on every row, which later can be merged with the transaction data to create a measure of centrality for each property.

import pandas as pd
import geopandas as gpd

# create dataframe containing geo coordinates of Trafalgar Square (to be trimmed later to match length of transaction dataset)

trafalgar_square = pd.DataFrame({'latitude': 51.5073509, 'longitude': -0.1277583}, index = range(350000))
trafalgar_square.head()

# convert to geo-dataframe, drop columns that are no longer needed, set CRS to Cartesian map coordinates

trafalgar_square_gdf = gpd.GeoDataFrame(trafalgar_square, geometry = gpd.points_from_xy(trafalgar_square['longitude'], trafalgar_square['latitude']))
trafalgar_square_gdf = trafalgar_square_gdf.drop(columns = ['latitude', 'longitude'])
trafalgar_square_gdf = trafalgar_square_gdf.set_crs(epsg=4326)
trafalgar_square_gdf = trafalgar_square_gdf.to_crs(epsg = 27700)
trafalgar_square_gdf.crs

trafalgar_square_gdf.to_file('../big_files/trafalgar_square_gdf.shp')
