==> 17_models_part_2.ipynb <==
**MODELLING PART 2**

So far the linear models (LR, Ridge, Lasso, ElasticNet) have performed very poorly compared to KNN and RandomForest. 

Performing a PowerTransformation on the target variable hasn't helped, it's possible that these models are struggling to capture a non-linear relationship between predictor variables and target.

A possible solution might be to try introducing Polynomial Features.


import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
sns.set_style('darkgrid')
from sklearn.linear_model import LinearRegression, Ridge, Lasso, ElasticNet, RidgeCV, LassoCV, ElasticNetCV
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import PolynomialFeatures, OneHotEncoder
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import r2_score
from sklearn.preprocessing import PowerTransformer
from sklearn.pipeline import Pipeline
from sklearn.compose import TransformedTargetRegressor

# import transaction dataframe, set relevant features to categorical dtype before dummifying

latest_df = pd.read_csv('../big_files/latest_df.csv')
latest_df['zone_50m'] = latest_df['zone_50m'].astype('category')
latest_df['zone_200m'] = latest_df['zone_200m'].astype('category')
latest_df['zone_400m'] = latest_df['zone_400m'].astype('category')
latest_df['zone_800m'] = latest_df['zone_800m'].astype('category')
latest_df['zone_1600m'] = latest_df['zone_1600m'].astype('category')
latest_df['zone_2400m'] = latest_df['zone_2400m'].astype('category')
latest_df['good_school'] = latest_df['good_school'].astype('category')
latest_df['tube_zone'] = latest_df['tube_zone'].astype('category')
latest_df['any_tube'] = latest_df['any_tube'].astype('category')
latest_df = latest_df.drop(columns = ['Unnamed: 0'])

# list of features used as predictor variables

updated_features = ['prop_type', 'fsm_lsoa', 'ea_in_ward', 'avg_airbnb', 'airbnb_tot', 'crime_lsoa', 'deli_count', 'flor_count', 'rest_count', 'income_rank_pos','employment_rank_pos', 'education_rank_pos', 'health_dep_score', 'crime_rank_pos', 'housing_rank_pos', 'living_env_pos', 'zone_50m','zone_200m', 'zone_400m', 'zone_800m', 'zone_1600m', 'zone_2400m', 'dist_traf', 'gang_prob_pos', 'crime_worry_pos', 'safety_fears_pos', 'satisfaction_pos', 'gun_crime', 'knife_crime', 'good_school', 'tube_zone', 'any_tube']
updated_df = latest_df[updated_features]
updated_df.head(1)

# create list of continous features for Polynomials (don't transform categorical variables)
# create a list of categorical features for dummifying

poly_features = [col for col in updated_features if updated_df[col].dtype == 'float' or latest_df[col].dtype == 'int']
cat_features = [col for col in updated_features if col not in poly_features]

# use ColumnTransformer to apply Polynomial transformation only to the continuous features, and dummify categorical features

poly = PolynomialFeatures(degree = 2)
ohe = OneHotEncoder()

t = [('polynomials', poly, poly_features), ('cat', ohe, cat_features)]

col_trans = ColumnTransformer(transformers = t, 
                              remainder='passthrough') 

updated_poly = col_trans.fit_transform(updated_df)

X = updated_poly
y = latest_df['price'].values

#train test split

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=1)

# standardize the data

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)


# create a linear regression model instance

model = LinearRegression(n_jobs = -2)

# get cross validated scores
scores = cross_val_score(model, X_train, y_train, cv=5)
print("Cross-validated training scores:", scores)
print("Mean cross-validated training score:", scores.mean())
# fit and evaluate the data on the whole training set
model.fit(X_train, y_train)
print("Training Score:", model.score(X_train, y_train))
# evaluate the data on the test set
print("Test Score:", model.score(X_test, y_test))

The Polynomial Features have given the LinearRegression model a significant boost, as might be expected.

But still far short of the performance of KNN and RandomForest.



# instantiate GridSearch object, and grid search for best parameters for Ridge model

ridge = Ridge()
params = {'alpha': np.logspace(-4, 4, 10)}
gs_ridge = GridSearchCV(ridge, params, cv = 5, n_jobs = -2, verbose = 3)
gs_ridge.fit(X_train, y_train)

# Ridge scores the same as Linear Regression (as it did without PolyNomial Features)

gs_ridge.best_score_

The PolyNomial Features have given a boost to the linear models.

We should try them with our most successful model, Random Forest.


rf = RandomForestRegressor(n_estimators = 500, verbose = 3, n_jobs = -2)
rf.fit(X_train, y_train)
cv_score = cross_val_score(rf, X_train, y_train, cv = 3)
test_predictions = rf.predict(X_test)
test_score = r2_score(y_test, test_predictions)


# the original RandomForest model (trained without PolyNomial Feautures) scored over 0.52, so in this case the PolyNomial Feature haven't helped.

print('mean cross_val_score', cv_score.mean())
print('test score', r2_score(y_test, test_predictions))

It seems unlikely that any of the linear models will get close to the scores achieved by Random Forest. 

However it might be worth trying Polynomial Features with interaction terms only.

# use ColumnTransformer to apply Polynomial transformation only to the continuous features, and dummify categorical features

poly = PolynomialFeatures(degree = 2, interaction_only = True)
ohe = OneHotEncoder()

poly_features = [col for col in updated_features if updated_df[col].dtype == 'float' or latest_df[col].dtype == 'int']
cat_features = [col for col in updated_features if col not in poly_features]

t = [('polynomials', poly, poly_features), ('cat', ohe, cat_features)]
col_trans = ColumnTransformer(transformers = t, 
                              remainder='passthrough') 

updated_poly = col_trans.fit_transform(updated_df)

X = updated_poly
y = latest_df['price'].values

#train test split

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=1)

# standardize the data

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

# create a linear regression model instance

model = LinearRegression(n_jobs = -2)

# get cross validated scores
scores = cross_val_score(model, X_train, y_train, cv=5)
print("Cross-validated training scores:", scores)
print("Mean cross-validated training score:", scores.mean())
# fit and evaluate the data on the whole training set
model.fit(X_train, y_train)
print("Training Score:", model.score(X_train, y_train))
# evaluate the data on the test set
print("Test Score:", model.score(X_test, y_test))

Interaction only performs slightly worse for Linear Regression than full Polynomial Features.

It might be worth trying it for Ridge too.


# instantiate ridgeCV model, and grid search for best parameter

ridge_cv = RidgeCV(alphas= np.logspace(-4, 4, 10), cv=5)
ridge_cv.fit(X_train, y_train)

ridge_cv.best_score_

The interaction only Polynomials haven't helped the linear models as much as the full Polynomial Features.

One other possibility is that we can revert to the full Polynomial Features and try transforming the target variable.

poly = PolynomialFeatures(degree = 2)
ohe = OneHotEncoder()

t = [('polynomials', poly, poly_features), ('cat', ohe, cat_features)]

col_trans = ColumnTransformer(transformers = t, 
                              remainder='passthrough') 

updated_poly = col_trans.fit_transform(updated_df)

X = updated_poly
y = latest_df['price'].values

#train test split

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=1)

# standardize the data

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

lr = LinearRegression()
pt = PowerTransformer(method = 'box-cox')

# PowerTransformer requires a matrix/DataFrame
pt.fit(y_train.reshape(-1, 1))
 
y_train_box = pt.transform(y_train.reshape(-1, 1))
y_test_box = pt.transform(y_test.reshape(-1, 1))

# fit the Linear Regression model on the transformed y_train, and make predictions

lr.fit(X_train, y_train_box)
predictions_trans = lr.predict(X_test)
predictions_trans

# transform the predictions back to the correct scale

real_predictions = pt.inverse_transform(predictions_trans)
real_predictions[:10]

# compare the predictions to y_test

r2_score(y_test, real_predictions)

The test score is lower than it was without the PowerTransformation.

It's unlikely to help, but we can try the same thing with other linear models.


# try the same thing with Ridge

ridge_cv_box = RidgeCV(alphas= np.logspace(-4, 4, 10), cv=5)
ridge_cv_box.fit(X_train, y_train_box)

ridge_box_preds = ridge_cv_box.predict(X_test)
ridge_box_preds

real_ridge_preds = pt.inverse_transform(ridge_box_preds)
r2_score(y_test, real_ridge_preds)

Experimenting with Polynomial Features has helped the linear models, as expected. Scores for the RandomForest haven't improved.

Combining the Polynomial Features with a transformation on the target variable didn't help.

One other possibilibity would be to take the Polynomial Features to a higher power (currently 2). However, this would explode the number of features, and given how far the linear models are lagging behind Random Forest it is unlikely to provide us with our best model.
