==> 21_data_dictionary.ipynb <==
import pandas as pd
latest_df = pd.read_csv('../../GA/big_files/latest_df.csv')

description_list = ['Sale price (£)', 'Date of sale', 'Postcode', 'Type of property', 'New build (y/n)', 'Estate type (freehold/leasehold)', 'Transaction type (Standard vs Other)', 'Year of sale', 'Month of sale', 'Lower Layer Super Output Area code', '% children eligible for free school meals in LSOA', 'Total estate agents in ward', 'Avg Airbnb nightly price for 2 bed flat n LSOA', 'Total volume of Airbnb listings in LSOA', 'Total crimes in LSOA', 'Total delicatessens in ward', 'Total florists in ward', 'Total top restaurants in ward', 'Department of Housing Income Rank (LSOA)', 'Department of Housing Employment Rank (LSOA)', 'Department of Housing Education Rank (LSOA)', 'Department of Housing Health Deprivation Score', 'Department of Housing Crime Rank', 'Department of Housing Accomodation Rank','Department of Housing Living Environment Rank', 'Nearest Tube within 50m (y/n)', 'Nearest Tube within 200m (y/n)', 'Nearest Tube within 400m (y/n)', 'Nearest Tube within 800m (y/n)', 'Nearest Tube within 1600m (y/n)', 'Nearest Tube within 2400m (y/n)', 'Distance to Trafalgar Square (metres)', 'Mayor Survey on gangs (% positive in borough)', 'Mayor Survey on crime (% positive in borough)', 'Mayor Survey on safety (% positive in borough)', 'Mayor Survey on satisfaction (% positive in borough)', 'Ward code', 'Ward name', 'Borough name', 'Volume of Gun Crime (ward)', 'Volume of Gun Crime (ward)', 'Ofsted Outstanding school within 500m (y/n)', 'Geo-coordinates', 'Tube Zone', 'Any tube within 2400m (y/n)']
sources = ['Land Registry', 'Land Registry', 'Land Registry', 'Land Registry', 'Land Registry', 'Land Registry', 'Land Registry', 'Land Registry', 'Land Registry', 'Greater London Authority', 'Greater London Authority', 'Scraped from All In London', 'Inside Airbnb', 'Inside Airbnb', 'Data.police.uk', 'Scraped from All In London', 'Scraped from All In London', 'Scraped from Good Food Guide', 'Ministry of Housing, Communities and Local Government', 'Ministry of Housing, Communities and Local Government', 'Ministry of Housing, Communities and Local Government', 'Ministry of Housing, Communities and Local Government', 'Ministry of Housing, Communities and Local Government', 'Ministry of Housing, Communities and Local Government', 'Ministry of Housing, Communities and Local Government',  'London Transport', 'London Transport', 'London Transport', 'London Transport', 'London Transport', 'London Transport', 'latitude.to', 'London Mayor Office for Policing and Crime', 'London Mayor Office for Policing and Crime', 'London Mayor Office for Policing and Crime', 'London Mayor Office for Policing and Crime', 'Greater London Authority', 'Greater London Authority', 'Greater London Authority', 'London Assembly', 'London Assembly', 'Ofsted', 'Derived from Postcode', 'London Transport', 'London Transport']
data_dict = pd.DataFrame({'variable_name': list(latest_df.columns)[1:], 'data_type': latest_df.dtypes.values[1:], 'description': description_list, 'source': sources})
data_dict
